requirements.txt
Django==2.2.3
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
